récupération et affichage des éléments de la BDD

+++ literie.sql
@@ -7,13 +7,13 @@ CREATE table catalogue (
     image varchar(255) not null,
     name varchar(100) not null,
     decription varchar(255),
-    price TINYINT
+    price decimal(6,2)
 );
 
 INSERT INTO catalogue 
 (image,name,decription,price)
 VALUES
-("almunda.jpeg","Almunda","Matelas Transition 90x190",759),
-("bjorn.jpeg","Bjorn","Matelas Stan 90x190",809),
-("eclipe.jpeg","Eclipse","Matelas Teamasse 140x190",759),
-("panama.jpeg","Panama","Matelas Coup de boule 160x200",1019)
+("almunda.jpeg","Almunda","Matelas Transition 90x190",759.00),
+("bjorn.jpeg","Bjorn","Matelas Stan 90x190",809.00),
+("eclipe.jpeg","Eclipse","Matelas Teamasse 140x190",759.00),
+("panama.jpeg","Panama","Matelas Coup de boule 160x200",1019.00)
